phpcs

+++ app/code/core/Mage/Usa/Model/Shipping/Carrier/Ups.php
@@ -1988,7 +1988,7 @@ XMLAuth;
             $successConversion = true;
             if ($responseCurrencyCode) {
                 if (in_array($responseCurrencyCode, $allowedCurrencies)) {
-                    $cost = (double)$cost * $this->_getBaseCurrencyRate($responseCurrencyCode);
+                    $cost = (float)$cost * $this->_getBaseCurrencyRate($responseCurrencyCode);
                 } else {
                     $errorTitle = __(
                         'We can\'t convert a rate from "%1-%2".',
